Fixed some glaring issues
Now object don't flicker behind walls, should've figured that one out ages ago.
Changed the coordinates, scale and rotation of things.
The underground parking garage is proper now, and the ground is generated properly to accommodate the garage.
the cars and containers are moved so they don't get in the way of the garage.
changed how the textures are put on the front of the mall