Changed some stuff
Added a drawBox method that draws the containers(the ones you specify the amount of at the start of the program, max being 9), changed the bmp file for the box.
Added a method that checks the numbers you enter at the start of the program, to check if it's under zero or over the max value.
The elevator doesn't clip through the walls anymore.